requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
openai
faiss-cpu
python-dotenv
pypdf